backend/requirements.txt
Django
djangorestframework
django-cors-headers
psycopg2
openai
numpy
scikit-learn
python-dotenv
django-filter
